feat: create selected element by dropping refactor: component splitting
