+++ src/components/blog/index.js
@@ -1,0 +1,5 @@
+import BlogUserInfo from './BlogUserInfo';
+import BlogFloatingUi from './BlogFloatingUi';
+import BlogGridLayout from './BlogGridLayout';
+
+export { BlogUserInfo, BlogFloatingUi, BlogGridLayout };
